requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
docutils==0.21.2
fastapi==0.114.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
imagesize==1.4.1
Jinja2==3.1.4
jiter==0.5.0
libsql-client==0.3.1
MarkupSafe==2.1.5
multidict==6.1.0
openai==1.45.0
packaging==24.1
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
starlette==0.38.5
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
